【修改内容】: 修改查询管网总长bug 【提交日期】: 20190911 【提交人】: fdk

+++ gis-service-beans/src/main/java/com/jdrx/gis/beans/entry/query/PipeLengthPO.java
@@ -13,7 +13,7 @@ import lombok.ToString;
 public class PipeLengthPO {
 
 	/** 总长度 */
-	private Integer totalLength;
+	private Double totalLength;
 
 	/** 日期 */
 	private String calDate;
